Added Java as an output generator and CircleCI job (#599)
* Added Java as an output generator and CircleCI job

* Getting closer (8/17): same file imports work

Java fixes:
* Files should be in a subdirectory reflected by the package path
* `Double`s can be null but `double`s cannot

* Standardized dictionary & set from index to get&set

Java uses .get and .put. Fine.

* Fixed HashSet's startNoItems; accepted baselines

* Missing ) in sortMembers.lambdaRight

* Missing required import for sets

* Lambda requiredImports and Java's .apply

* Accepted integration baselines

* Java does not include index in string formatting. 12/17 passing!

* Java use charAt for string index, not []

* Use add() for Java list auto initializers

`List.of` is immutable, unfortunately.

* Static generic functions in Java put the generic before the return

* Java uses .get() because OF COURSE

* Comparator.naturalOrder and no ternaries for Java ArrayList sorts

* Generic types need to also add type imports

* Added a list set command

Java uses `.set()` instead of `[]`.

* Added ArrayNewGeneric, ArrayNewSizedGeneric commands

Java needs them to make a new `(T[]) Object[]` array initializer

* Removed warnings from javac

* Implemented { throws } for member functions and main context

* A few more baselines...

* Deep wildcard for Java file paths

* Fibonacci: (I)Generator.gls; int instead of number type

* Reset and accepted new Fibonacci baselines

* Fixed a few more merge conflict artifacts

* Simplified JavaOutputGenerator a bit more

* ...and a bit more

+++ test/end-to-end/Fibonacci/C#/Main.cs
@@ -6,7 +6,7 @@ namespace Fibonacci
 
     class Program
     {
-        private static string GetLabel(double index)
+        private static string GetLabel(int index)
         {
             if (index == 1)
             {
